Ready first beta version

+++ tcreator/src/main/java/com/lvg/tcreator/services/GenericService.java
@@ -3,7 +3,7 @@ package com.lvg.tcreator.services;
 public interface GenericService<T> {
 	
 	public void save(T record);
-	public T get(int id);
+	public T get(long id);
 	public void update(T record);
 	public void delete(T record);
